requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
python-dotenv
PyMySQL
aiomysql
pytest
httpx